Update to the latest what4 changes
This includes changes to the BaseFloatType operations and a change to the
signature of symbolic functions.  Some of the serialization code required new
constraints.  The BaseNatType has been demoted and is no longer a base type (it
was not soundly handled).  This code did not really use the Nat type, so the
impact was minimal.