fix 'setSlice' ASL override when setting high bits
The ASL bitvector slicing primitives are pre-processed
into calls to 'setSlice' and 'getSlice' respectively to
implement bitvector slicing in terms of bitvector masking
and shifting.

This fixes an error in the 'setSlice' implementation which
affected the case where a smaller bitvector is assigned to the
high bits of another bitvector. This affected multiple instructions,
including 'MOVT', where the instruction would simply do nothing instead
of assigning the correct value to the target register.

Includes updated archived semantics as a result of re-running
the translator after applying the patch.