Bump what4 submodule to version 1.4
This version of `what4` includes the changes from GaloisInc/what4#217, which
merges the `what4-serialize` library into `what4`. As a result, some of the
imports in `asl-translator` had to be adjusted somewhat.

Unlike `what4-serialize`, `what4` does not define a `SomeSome` data type, so I
introduce this in the `Data.Parameterized.SomeSome` module.