Embed conditions into uf_read_mem to detect spurious reads
This is in support of: https://github.com/GaloisInc/macaw/issues/266

Adds an additional step to term normalization that
changes calls to uf_read_mem_# (representing memory reads)
into uf_cond_read_mem_#, which has the most recent
Mux condition added as an argument.

This allows macaw to check this condition at runtime,
and avoid emitting a spurious memory read if it
is concretely false.